feat(util): convert leaf utilities to TypeScript
- Convert constants.js → constants.ts (named exports)
- Convert timer.js → timer.ts (Timer class + utility functions)
- Convert decay.js → decay.ts (typed decay calculations)
- Convert languages.js → languages.ts (language map)
- Convert keywords.js → keywords.ts (typed keyword helpers)
- Convert scrape.js → scrape.ts (typed scraping functions)
- Convert trackerInputParser.js → trackerInputParser.ts (typed parser)

Phase 2 complete.
