--- src/util/scrape.js
@@ -1,82 +1,0 @@
-const Axios = require('axios');
-
-const Constants = require('../util/constants');
-const Utils = require('../util/utils');
-
-module.exports = {
-    scrape: async function (url) {
-        try {
-            return await Axios.get(url);
-        } catch (e) {
-            return {};
-        }
-    },
-
-    scrapeSteamProfilePicture: async function (client, steamId) {
-        const response = await module.exports.scrape(`${Constants.STEAM_PROFILES_URL}${steamId}`);
-
-        if (response.status !== 200) {
-            client.log(
-                client.intlGet(null, 'errorCap'),
-                client.intlGet(null, 'failedToScrapeProfilePicture', {
-                    link: `${Constants.STEAM_PROFILES_URL}${steamId}`,
-                }),
-                'error',
-            );
-            return null;
-        }
-
-        let png = response.data.match(/<img src="(.*_full.jpg)(.*?(?="))/);
-        if (png) {
-            return png[1];
-        }
-
-        return null;
-    },
-
-    scrapeSteamProfileName: async function (client, steamId) {
-        const response = await module.exports.scrape(`${Constants.STEAM_PROFILES_URL}${steamId}`);
-
-        if (response.status !== 200) {
-            client.log(
-                client.intlGet(null, 'errorCap'),
-                client.intlGet(null, 'failedToScrapeProfileName', {
-                    link: `${Constants.STEAM_PROFILES_URL}${steamId}`,
-                }),
-                'error',
-            );
-            return null;
-        }
-
-        let regex = new RegExp(`class="actual_persona_name">(.+?)</span>`, 'gm');
-        let data = regex.exec(response.data);
-        if (data) {
-            return Utils.decodeHtml(data[1]);
-        }
-
-        return null;
-    },
-
-    scrapeSteamIdFromVanity: async function (client, vanity) {
-        if (typeof vanity !== 'string' || vanity.trim() === '') return null;
-
-        const url = `https://steamcommunity.com/id/${encodeURIComponent(vanity.trim())}/?xml=1`;
-        const response = await module.exports.scrape(url);
-
-        if (response.status !== 200 || typeof response.data !== 'string') {
-            return null;
-        }
-
-        const steamId64Match = response.data.match(/<steamID64>(\d{17})<\/steamID64>/i);
-        if (steamId64Match) {
-            return steamId64Match[1];
-        }
-
-        const profileLinkMatch = response.data.match(/steamcommunity\.com\/profiles\/(\d{17})/i);
-        if (profileLinkMatch) {
-            return profileLinkMatch[1];
-        }
-
-        return null;
-    },
-};
